+++ java/client/src/main/java/module-info.java
@@ -4,6 +4,7 @@ module glide.api {
     exports glide.api.logging;
     exports glide.api.models;
     exports glide.api.models.commands;
+    exports glide.api.models.commands.batch;
     exports glide.api.models.commands.bitmap;
     exports glide.api.models.commands.geospatial;
     exports glide.api.models.commands.function;
@@ -15,13 +16,7 @@ module glide.api {
     exports glide.api.models.exceptions;
     exports glide.api.commands.servermodules;
 
-    requires com.google.protobuf;
-    requires io.netty.codec;
-    requires io.netty.common;
-    requires io.netty.transport;
-    requires io.netty.transport.classes.epoll;
-    requires io.netty.transport.classes.kqueue;
-    requires io.netty.transport.unix.common;
-    requires lombok;
+    requires java.logging; // required by shadowed protobuf
+    requires static lombok;
     requires org.apache.commons.lang3;
 }