+++ java/integTest/src/test/java/glide/TestConfiguration.java
@@ -4,12 +4,12 @@ package glide;
 import static glide.TestUtilities.commonClientConfig;
 import static glide.TestUtilities.commonClusterClientConfig;
 
-import com.vdurmont.semver4j.Semver;
 import glide.api.BaseClient;
 import glide.api.GlideClient;
 import glide.api.GlideClusterClient;
 import glide.api.logging.Logger;
 import org.apache.commons.lang3.tuple.Pair;
+import org.semver4j.Semver;
 
 public final class TestConfiguration {
     public static final String[] STANDALONE_HOSTS =