+++ java/integTest/src/test/java/glide/TestUtilities.java
@@ -32,6 +32,7 @@ import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import lombok.NonNull;
 import lombok.SneakyThrows;
 import lombok.experimental.UtilityClass;
@@ -89,6 +90,19 @@ public class TestUtilities {
                                 parts -> parts[1],
                                 (existingValue, newValue) -> newValue,
                                 LinkedHashMap::new));
+    }
+
+    // copied from glide.utils.ArrayTransformUtils.concatenateArrays, because it is not exported
+    /**
+     * Concatenates multiple arrays of type T and returns a single concatenated array.
+     *
+     * @param arrays Varargs parameter for arrays to be concatenated.
+     * @param <T> The type of the elements in the arrays.
+     * @return A concatenated array of type T.
+     */
+    @SafeVarargs
+    public static <T> T[] concatenateArrays(T[]... arrays) {
+        return Stream.of(arrays).flatMap(Stream::of).toArray(size -> Arrays.copyOf(arrays[0], size));
     }
 
     public static GlideClientConfiguration.GlideClientConfigurationBuilder<?, ?>